pyproject.toml
[project]
name = "olt-provisioning-api"
version = "1.0.0"
description = "JTL ZTE OLT Provisioning API Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "aiomysql>=0.2",
    "alembic>=1.14",
    "asyncssh>=2.17",
    "pydantic-settings>=2.6",
    "cryptography>=44.0",
    "structlog>=24.4",
    "africastalking>=1.2",
    "aiosmtplib>=3.0",
    "jinja2>=3.1",
    "pymysql>=1.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
